Protect category routes behind auth, allow only RTPs and contributors to delete and create categories

+++ ctf/routes/categories.py
@@ -2,14 +2,17 @@
 
 Contains the routes pertaining to the categories a challenge can fit in to
 """
-from flask import Blueprint, jsonify, request
+from flask import Blueprint, jsonify, request, session
 
+from ctf import auth
 from ctf.models import Categories
+from ctf.ldap import is_ctf_admin
 
 categories_bp = Blueprint('categories_bp', __name__)
 
 
 @categories_bp.route('/', methods=['GET', 'POST'])
+@auth.oidc_auth
 def all_categories():
     """
     Operations relating categories
@@ -18,9 +21,15 @@ def all_categories():
     :POST: Takes 'name' and 'description' values from application/json body and creates a new
         category
     """
+    print(session['userinfo'].get('preferred_username'))
     if request.method == 'GET':
         return jsonify([category.to_dict() for category in Categories.query.all()]), 200
     elif request.method == 'POST':
+        if not is_ctf_admin(session['userinfo'].get('preferred_username')):
+            return jsonify({
+                'status': "error",
+                'message': "You aren't authorized to create categories"
+            }), 403
         if not request.is_json:
             return jsonify({
                 'status': "error",
@@ -42,6 +51,7 @@ def all_categories():
 
 
 @categories_bp.route('/<category_name>', methods=['GET', 'DELETE'])
+@auth.oidc_auth
 def single_category(category_name: str):
     """
     Operations relating to a single category
@@ -60,5 +70,10 @@ def single_category(category_name: str):
     if request.method == 'GET':
         return jsonify(category.to_dict()), 200
     elif request.method == 'DELETE':
+        if not is_ctf_admin(session['userinfo'].get('preferred_username')):
+            return jsonify({
+                'status': "error",
+                'message': "You aren't authorized to create categories"
+            }), 403
         category.delete()
         return '', 204
